requirements.txt
playwright
pytz
python-dateutil
sqlite3
google-generativeai
pydantic
pystray
pillow
olefile